# Performance fix: avoid unnecessary vector copy